requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
apispec==6.8.1
asgiref==3.8.1
asyncer==0.0.7
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bech32==1.2.0
bidict==0.23.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
chainlit==1.3.2
charset-normalizer==3.4.0
chevron==0.14.0
chroma-hnswlib==0.7.6
chromadb==0.5.20
click==8.1.7
coloredlogs==15.0.1
cosmpy==0.9.2
dataclasses-json==0.6.7
Deprecated==1.2.15
durationpy==0.9
ecdsa==0.19.0
fastapi==0.115.5
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
gpt4all==2.8.2
greenlet==3.1.1
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
langchain==0.3.9
langchain-community==0.3.8
langchain-core==0.3.29
langchain-ollama==0.2.2
langchain-text-splitters==0.3.2
langchainhub==0.1.21
langsmith==0.1.147
Lazify==0.4.0
literalai==0.0.623
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.5
onnxruntime==1.20.1
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.12
overrides==7.7.0
packaging==23.2
posthog==3.7.4
propcache==0.2.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.10.5
pydantic-settings==2.6.1
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.9
python-socketio==5.12.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
setuptools==75.6.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.41.3
sympy==1.13.3
syncer==2.0.3
tenacity==9.0.0
tokenizers==0.21.0
tomli==2.2.1
tqdm==4.67.1
typer==0.14.0
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
uagents==0.18.1
uptrace==1.28.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==0.20.0
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.0
zipp==3.21.0
